Use  'multinom()' instead of 'mlogit()' in bootstrap.
Correct the problem associated to the incorrect use of 'sort()' for the mediator levels in 'exactmed_cat()'.

Change the mediator levels in 'datamed_cat' dataframe.

Add check for the values of the input parameters 'a0' and'a1'.

Add the value of the mediator in controlled effects output for continuous and for categorical versions.

Add tests for 'exactmed_cat()' and for the new check associated to 'a0' and 'a1'parameters.

Add a specific section for mediation analysis with a categorical mediator in vignette.

Apply minor corrections in description, functions help, test and vignette files.